backend/requirements.txt
djangorestframework
mysqlclient
django-cors-headers
djangorestframework-simplejwt
timedelta
pandas